trait IntoParserResult ->IntoParserResultAddSpan and add new IntoParserResultAddCode
This cannot be one trait. They are implemented for types that cannot implement the second method for lack of data.

The first is for converting external error types that lack the span. The second for internal error chaining when there already is a useful span and changing that makes little sense.